Move make_test macros to a separate utils ns
